# docs/openapi.yaml
openapi: 3.0.0
info:
  version: 1.0.0
  title: nodejs_hw_07
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  description: >
    Rehber API uygulaması - Node.js kursunun swagger dokümantasyon ödevi
tags:
  - name: Contacts
    description: Kişi rehberi işlemleri
servers:
  - url: http://localhost:3000
  #- url: <Render.com URL'iniz>
paths:
  /contacts:
    get:
      tags:
        - Contacts
      summary: Tüm kişileri listeleme
      operationId: getAllContacts
      description: Sistemde kayıtlı tüm kişileri listeler
      security:
        - bearerAuth: []
      parameters:
        - name: page
          in: query
          required: false
          description: Sayfa numarası (varsayılan 1)
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          description: Sayfa başına kayıt sayısı (varsayılan 10)
          schema:
            type: integer
            default: 10
      responses:
        '200':
          description: Başarılı
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Contact'
    post:
      tags:
        - Contacts
      summary: Yeni kişi ekleme
      operationId: createContact
      description: Sisteme yeni bir kişi kaydı ekler
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - name
              properties:
                name:
                  type: string
                  description: Kişinin adı soyadı
                email:
                  type: string
                  format: email
                  description: Kişinin e-posta adresi
                phone:
                  type: string
                  description: Kişinin telefon numarası
      responses:
        '201':
          description: Kayıt başarılı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
  /contacts/{contactId}:
    get:
      tags:
        - Contacts
      summary: Belirli bir kişiyi getirme
      operationId: getContactById
      description: ID'ye göre belirli bir kişi kaydını getirir.
      security:
        - bearerAuth: []
      parameters:
        - name: contactId
          in: path
          required: true
          description: Getirilecek kişinin ID'si
          schema:
            type: string
      responses:
        '200':
          description: Başarılı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags:
        - Contacts
      summary: Kişi bilgilerini güncelleme
      operationId: updateContact
      description: Belirli bir kişinin bilgilerini günceller
      security:
        - bearerAuth: []
      parameters:
        - name: contactId
          in: path
          required: true
          description: Güncellenecek kişinin ID'si
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                  description: Kişinin yeni adı soyadı
                email:
                  type: string
                  format: email
                  description: Kişinin yeni e-posta adresi
                phone:
                  type: string
                  description: Kişinin yeni telefon numarası
      responses:
        '200':
          description: Başarılı
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Contact'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      tags:
        - Contacts
      summary: Kişi silme
      operationId: deleteContact
      description: Belirli bir kişiyi sistemden siler
      security:
        - bearerAuth: []
      parameters:
        - name: contactId
          in: path
          required: true
          description: Silinecek kişinin ID'si
          schema:
            type: string
      responses:
        '204':
          description: Silme işlemi başarılı (içerik yok)
        '404':
          $ref: '#/components/responses/NotFound'
components:
  schemas:
    Contact:
      type: object
      required:
        - id
        - name
      properties:
        id:
          type: string
          description: Kişinin benzersiz kimliği
        name:
          type: string
          description: Kişinin adı soyadı
        email:
          type: string
          format: email
          description: Kişinin e-posta adresi
        phone:
          type: string
          description: Kişinin telefon numarası
  responses:
    NotFound:
      description: Kaynak bulunamadı
      content:
        application/json:
          schema:
            type: object
            properties:
              message:
                type: string
                example: "Kişi bulunamadı"
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer